Fixed refactoring bug
