--- src/main/java/com/cgm/infolab/db/model/enumeration/RoomTypeEnum-N-CBD6YT2.java
@@ -1,5 +1,0 @@
-package com.cgm.infolab.db.model.enumeration;
-
-public enum RoomTypeEnum {
-    GROUP, USER2USER
-}
