+++ src/main/java/com/cgm/infolab/db/model/enumeration/RoomTypeEnum.java
@@ -1,4 +1,4 @@
-package com.cgm.infolab.db.model;
+package com.cgm.infolab.db.model.enumeration;
 
 public enum RoomTypeEnum {
     GROUP, USER2USER
